Move command back into root of commands folder
