+++ src/commands/deleteFunctionApp.ts
@@ -4,9 +4,9 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { IActionContext } from '@microsoft/vscode-azext-utils';
-import { ext } from '../../extensionVariables';
-import { ResolvedFunctionAppResource } from '../../tree/ResolvedFunctionAppResource';
-import { SlotTreeItem } from '../../tree/SlotTreeItem';
+import { ext } from '../extensionVariables';
+import { ResolvedFunctionAppResource } from '../tree/ResolvedFunctionAppResource';
+import { SlotTreeItem } from '../tree/SlotTreeItem';
 
 export async function deleteFunctionApp(context: IActionContext, node?: SlotTreeItem): Promise<void> {
     if (!node) {
